# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "User Validation API",
    "description": "REST API for creating and fetching users",
    "contact": {
      "name": "Spring Validation",
      "email": "support@example.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9292",
      "description": "Local development server"
    }
  ],
  "tags": [
    {
      "name": "Users",
      "description": "User management APIs"
    }
  ],
  "paths": {
    "/users/add": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a user",
        "description": "Adds a new user to the database",
        "operationId": "saveUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/users/user/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by ID",
        "description": "Returns a single user for the given ID",
        "operationId": "getAllUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/users/fetchAll": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users",
        "description": "Returns every user stored in the database",
        "operationId": "getAllUser_1",
        "responses": {
          "200": {
            "description": "Users fetched successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserRequest": {
        "type": "object",
        "description": "Request payload for creating a user",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "pankhuri"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "pankhuri@gmail.com"
          },
          "mobile": {
            "type": "string",
            "description": "Mobile number",
            "example": "7423456789"
          },
          "gender": {
            "type": "string",
            "description": "Gender",
            "example": "Female"
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "description": "Age in years",
            "example": 27
          },
          "state": {
            "type": "string",
            "description": "State of residence",
            "example": "Bihar"
          }
        }
      },
      "User": {
        "type": "object",
        "description": "User entity stored in the database",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32",
            "description": "Auto-generated user ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "pankhuri"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "pankhuri@gmail.com"
          },
          "mobile": {
            "type": "string",
            "description": "Mobile number",
            "example": "7423456789"
          },
          "gender": {
            "type": "string",
            "description": "Gender",
            "example": "Female"
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "description": "Age in years",
            "example": 27
          },
          "state": {
            "type": "string",
            "description": "State of residence",
            "example": "Bihar"
          }
        }
      }
    }
  }
}