# src/main/java/com/spring_validation/validation/controller/UserController.java
package com.spring_validation.validation.controller;

import com.spring_validation.validation.dto.UserRequest;
import com.spring_validation.validation.entity.User;
import com.spring_validation.validation.service.UserServiceImpl;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/users")
@Tag(name = "Users", description = "User management APIs")
public class UserController {
    @Autowired
    private UserServiceImpl userService;

    @PostMapping("/add")
    @Operation(summary = "Create a user", description = "Adds a new user to the database")
    @ApiResponses({
            @ApiResponse(responseCode = "201", description = "User created successfully",
                    content = @Content(schema = @Schema(implementation = User.class))),
            @ApiResponse(responseCode = "400", description = "Invalid request payload"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
    })
    public ResponseEntity<User> saveUser(@RequestBody UserRequest userRequest) {
        return new ResponseEntity<>(userService.saveUser(userRequest), HttpStatus.CREATED);
    }

    @GetMapping("/fetchAll")
    @Operation(summary = "Get all users", description = "Returns every user stored in the database")
    @ApiResponse(responseCode = "200", description = "Users fetched successfully")
    public ResponseEntity<List<User>> getAllUser() {
        return ResponseEntity.ok(userService.getAllUser());
    }

    @GetMapping("/user/{id}")
    @Operation(summary = "Get user by ID", description = "Returns a single user for the given ID")
    @ApiResponses({
            @ApiResponse(responseCode = "200", description = "User found",
                    content = @Content(schema = @Schema(implementation = User.class))),
            @ApiResponse(responseCode = "404", description = "User not found")
    })
    public ResponseEntity<User> getAllUser(
            @Parameter(description = "User ID", example = "1") @PathVariable int id) {
        return ResponseEntity.ok(userService.getAllUser(id));
    }
}