# src/main/java/com/spring_validation/validation/dto/UserRequest.java
package com.spring_validation.validation.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.*;

@Data
@Getter
@Setter
@AllArgsConstructor(staticName = "build")
@NoArgsConstructor
@Schema(description = "Request payload for creating a user")
public class UserRequest {
    @Schema(description = "Full name of the user", example = "pankhuri")
    private String name;

    @Schema(description = "Email address", example = "pankhuri@gmail.com")
    private String email;

    @Schema(description = "Mobile number", example = "7423456789")
    private String mobile;

    @Schema(description = "Gender", example = "Female")
    private String gender;

    @Schema(description = "Age in years", example = "27")
    private int age;

    @Schema(description = "State of residence", example = "Bihar")
    private String state;
}